// src/app/client/model/presets.ts
/**
 * ViewPI Camera
 * Backend for RaspiMjpeg
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Presets { 
    id?: number | null;
    mode: string;
    name: string;
    width: number;
    height: number;
    fps: number;
    i_width: number;
    i_height: number;
    i_rate: number;
}